requirements.txt
# WeChef — dépendances Python
# Généré le 2026-03-13

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.115
uvicorn[standard]>=0.29
python-multipart>=0.0.9      # upload fichiers

# ── Base de données ───────────────────────────────────────────────────────────
sqlalchemy>=2.0
alembic>=1.13                 # migrations
asyncpg>=0.29                 # driver PostgreSQL async (prod)
aiofiles>=23.2                # lecture fichiers async

# ── Validation ────────────────────────────────────────────────────────────────
pydantic>=2.7
pydantic-settings>=2.3
python-dotenv>=1.0

# ── LLM / IA ─────────────────────────────────────────────────────────────────
google-genai>=0.8

# ── Extraction vidéo ─────────────────────────────────────────────────────────
# yt-dlp est installé système (/usr/bin/yt-dlp) via apt ou pip global
# yt-dlp>=2024.1

# ── Images ───────────────────────────────────────────────────────────────────
httpx>=0.27

# ── PDF ───────────────────────────────────────────────────────────────────────
weasyprint>=62

# ── Tests ────────────────────────────────────────────────────────────────────
pytest>=8.2
pytest-asyncio>=0.23
httpx>=0.27                   # TestClient async
anyio[trio]>=4.4
